requirements.txt
requests
kafka-python
urllib3==1.26.15
boto3
